renamed kuzzleMemoryStorage.js -> MemoryStorage.js
